content-meme-automation/requirements.txt
# LangGraph Content & Meme Generation Automation

## Core Dependencies
langgraph>=0.2.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

## Document Parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
markdown>=3.5.0

## Image Processing
Pillow>=10.0.0
rembg>=2.0.50
opencv-python>=4.8.0

## Google Cloud (for Veo 3, Imagen)
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0

## Web Scraping & APIs
tweepy>=4.14.0
snscrape>=0.6.2
praw>=7.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0

## Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

## CLI
click>=8.1.0
rich>=13.7.0
